requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
psycopg2-binary==2.9.6
pandas==2.0.3
scikit-learn==1.3.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
google-generativeai==0.3.0
sib-api-v3-sdk==7.6.0
python-dotenv==1.0.0
itsdangerous==2.1.2